Implements Flow for User Circuit Registration for Worker (#12)
* feat: prover registration flow with few shallow functions


* fix: update task status to in progress too


---------

+++ quantum_types/src/types/db/user_circuit_data.rs
@@ -7,7 +7,7 @@ pub struct UserCircuitData{
     pub circuit_hash: String,
     pub vk_path: String,
     pub reduction_circuit_id: Option<u64>,
-    pub pis_len: u64,
+    pub pis_len: u8,
     pub proving_scheme: ProvingSchemes,
     pub circuit_reduction_status: CircuitReductionStatus
 }
